requirements.txt
# TaxLens-AI — Python 3.10+
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pandas>=2.0.0
openpyxl>=3.1.0
numpy>=1.26.0

# Local RAG (Ollama must be running on-premise; no cloud)
llama-index-core>=0.10.30
llama-index-llms-ollama>=0.1.3
llama-index-embeddings-ollama>=0.1.2

# Optional OCR / PDF (install when enabling on-premise OCR pipeline)
# pymupdf>=1.23.0
# pytesseract>=0.3.10
# Pillow>=10.0.0
